requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.2.0
google-api-core==2.11.1
google-auth==2.21.0
google-cloud==0.34.0
google-cloud-texttospeech==2.14.1
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
idna==3.4
lxml==4.9.3
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
soupsieve==2.4.1
urllib3==1.26.16
